Updated the README.md. Added the Reset button and a text box with the money available to the ControlPanel. Small code cleanups.
