+++ game/configs/tileset_settings.yaml
@@ -5,7 +5,6 @@ tile_height: 8
 tiles:
   grass:
     main: [[7,11]]
-#  clay: [[7,10]]
   cobblestone:
     main: [[3,16]]
     money: 0
